requirements.txt
torch
torchvision
opencv-python
numpy
pandas
scikit-learn
streamlit
Pillow
